Fetchdata/requirements.txt
brotli==1.0.9
brotlicffi==1.0.9.2
chardet==5.1.0
ConfigParser==5.3.0
cryptography==40.0.1
mysql_connector_python==8.0.32
protobuf==4.22.3
pyOpenSSL==23.1.1
pytz==2022.7.1
simplejson==3.19.1
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
